feat: implements toggle with plan labels
